+++ src/compas_cem/optimization/parameters/edge.py
@@ -24,27 +24,40 @@ class EdgeParameter(Serializable):
 
     def key(self):
         """
+        The edge key.
         """
         return self._key
 
     def start_value(self, topology):
         """
+        The initial value of the edge optimization parameter.
         """
         val = topology.edge_attribute(key=self.key(), name=self._attr_name)
         return val
 
     def bound_low(self, topology):
         """
+        The lower bound of the edge optimization parameter.
+
+        Notes
+        -----
+        Calculated as the initial parameter minus this bound's absolute value.
         """
         return self.start_value(topology) - fabs(self._bound_low)
 
     def bound_up(self, topology):
         """
+        The upper bound of the edge optimization parameter.
+
+        Notes
+        -----
+        Calculated as the initial parameter plus this bound's absolute value.
         """
         return self.start_value(topology) + fabs(self._bound_up)
 
     def attr_name(self):
         """
+        The name of the edge attribute to parametrize.
         """
         return self._attr_name